+++ backend/prisma/migrations/20260223084642_adjust_membershp/migration.sql
@@ -1,0 +1,9 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `note` on the `Membership` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE "Membership" DROP COLUMN "note",
+ADD COLUMN     "point" INTEGER NOT NULL DEFAULT 0;